Play nicer with font-lock
Since this makes the switch to using 'font-lock-face instead of 'face text
property (so that font-lock doesn't touch it when refontifying), the
highlighting now depends on font-lock being enabled.

I suppose we could check whether font-lock is enabled at runtime and use the one
or the other property, but really, who would disable font-lock and expect the
syntax highlighting to work?

Refs #32.