Classify function variables traversing its AST just once
This is one step further to separate the new functionality from the
js2 core, as the highlighter function does not rely on the global
js2-recorded-identifier. It should be easier now exposing it as a
minor mode as suggested by Dmitry.

Now all the uses of uninitialized variables are highlighted, not their
declaration.