Add js2-declare-variable
js2-declare-variable makes it easy to extend var statements.  Bound to
C-c C-v, just enter the name of a variable (and optionally its initial
value) and js2-declare-variable will append the declaration to the first
var statement in the scope enclosing point.

If there isn't already a var statement in the enclosing scope, one will
be created.

If there is a "use strict" statement or comments at the top of the
enclosing scope, the var statement will be inserted after them.