Fix #356 by commenting map entries for js2-mode-show-node
js2-mode-show-node is a debugging aid. My elisp skills are weak and I
could not find what, within that function, causes #356. So I just
commented out the two references to it, and maybe the js2-mode maintainer
has a good idea of the best way to workaround the bad behavior on
macOS/OSX.