Indent arrays correctly when they contain for loops.
The "js2-indent-in-array-comp" function attempts to detect array
comprehensions, which are described on MDN here:

https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Array_comprehensions

It is mildly over-eager, and causes the following code to be indented
like so:

var x = [
  function () {
    for (var i; i < 10; i++) {
    }
  },
    ];

With the closing bracket of the array indented to the same level as
the `for' statement that happens to be inside the array.

This patch helps prevent that, by checking that the paren depth of the
for loop is exactly one plus the paren depth of the containing array.